api: move the peer-lookup budget into its own file
pkg/api/authz.go sat one line under the 1500 LOC modularity floor, so the
second pool and its documentation crossed it — for a reason that has
nothing to do with authz.go's own subject.

The budget is a cohesive unit on its own terms, not a convenient slice: a
bound, the predicate that decides which bound applies to a connection, and
the observability for both. Moved whole, with no behaviour change; the
file header records why it is separate so the next reader does not fold it
back in.

Validation: pkg/api, pkg/authz and pkg/refactoraudit green.

Refs #6974
